Issue #121: Move Address-related stuff out of wallet_model
* Moved AddressRecord and LooseAddressRecord into address.py
* Moved DeterministicAddressRecord and DWalletAddressManager into deterministic.py
* Moved BIP0032AddressRecord and HDWalletAddressManager into bip0032.py
* Added unit tests for all 3
* Fixed a couple bugs related to importing loose addresses
