+++ ngcccbase/wallet_model.py
@@ -8,466 +8,18 @@ definitions, but it doesn't implement high-level operations
 (those are implemented in controller).
 """
 
-import hmac
-import hashlib
-import json
-import os
-
 from collections import defaultdict
-from pycoin.ecdsa import Point
-from pycoin.ecdsa.secp256k1 import generator_secp256k1 as BasePoint
-from pycoin.encoding import (b2a_base58, from_bytes_32,
-                             hash160_sec_to_bitcoin_address,
-                             public_pair_to_bitcoin_address)
-from pycoin.wallet import Wallet
+from pycoin.encoding import hash160_sec_to_bitcoin_address
 
-from asset import AssetDefinition, AssetDefinitionManager
+from asset import AssetDefinitionManager
 from coloredcoinlib import (ColorSet, BlockchainState, ColorDataBuilderManager,
                             FullScanColorDataBuilder, DataStoreConnection,
                             ColorDataStore, ColorMetaStore, ColorMap,
                             ThickColorData, toposorted)
-from ngcccbase.services.electrum import EnhancedBlockchainState
-from ngcccbase import txdb
+from services.electrum import EnhancedBlockchainState
+from txdb import TxDb
 from txcons import TransactionSpecTransformer
 from utxodb import UTXOQuery, UTXOManager
-
-
-class AddressRecord(object):
-    """Object that holds both address and color information
-    Note this is now an Abstract Class.
-    """
-    def __init__(self, **kwargs):
-        pass
-
-    def get_color_set(self):
-        """Access method for the color set associated
-        with this address record
-        """
-        return self.color_set
-
-    def get_data(self):
-        """Get this object as a JSON/Storage compatible dict.
-        Useful for storage and persistence.
-        """
-        return {"color_set": self.color_set.get_data(),
-                "address_data": b2a_base58(to_byte_32(self.rawPrivKey))}
-
-    def get_address(self):
-        """Get the actual bitcoin address
-        """
-        return self.address
-
-    def get_color_address(self):
-        """This is the address that can be used for sending/receiving
-        colored coins
-        """
-        if self.color_set.uncolored_only():
-            return self.get_address()
-        return "%s@%s" % (self.get_color_set().get_color_hash(),
-                          self.get_address())
-
-
-class DeterministicAddressRecord(AddressRecord):
-    """Subclass of AddressRecord which is entirely deterministic.
-    DeterministicAddressRecord will use a single master key to
-    create addresses for specific colors and bitcoin addresses.
-    """
-    def __init__(self, **kwargs):
-        """Create an address for this color <color_set>
-        and index <index> with the master key <master_key>.
-        The address record returned for the same three variables
-        will be the same every time, hence "deterministic".
-        """
-        self.color_set = kwargs.get('color_set')
-        if len(self.color_set.get_data()) == 0:
-            color_string = "genesis block"
-        else:
-            color_string = self.color_set.get_hash_string()
-
-        self.testnet = kwargs.get('testnet')
-        h = hmac.new(str(kwargs['master_key']),
-                     "%s|%s" % (color_string, kwargs['index']), hashlib.sha256)
-        string = h.digest()
-        self.rawPrivKey = from_bytes_32(string)
-        self.publicPoint = BasePoint * self.rawPrivKey
-        self.address = public_pair_to_bitcoin_address(self.publicPoint.pair(),
-                                                      compressed=False,
-                                                      is_test=self.testnet)
-
-
-class BIP0032AddressRecord(AddressRecord):
-    """Subclass of AddressRecord which is deterministic and BIP0032 compliant.
-    BIP0032AddressRecord will use a pycoin wallet to create addresses
-    for specific colors.
-    """
-    def __init__(self, **kwargs):
-        """Create an address for this color <color_set> and index <index>
-        with the pycoin_wallet <pycoin_wallet> and on testnet or not
-        <testnet>
-        The address record returned for the same variables
-        will be the same every time, hence "deterministic".
-        """
-        pycoin_wallet = kwargs.get('pycoin_wallet')
-        self.color_set = kwargs.get('color_set')
-        color_string = hashlib.sha256(self.color_set.get_earliest()).digest()
-
-        # use the hash of the color string to get the subkey we need
-        while len(color_string):
-            number = int(color_string[:4].encode('hex'), 16)
-            pycoin_wallet = pycoin_wallet.subkey(i=number, is_prime=True,
-                                                 as_private=True)
-            color_string = color_string[4:]
-
-        # now get the nth address in this wallet
-        pycoin_wallet = pycoin_wallet.subkey(i=kwargs.get('index'),
-                                             is_prime=True, as_private=True)
-
-        self.testnet = kwargs.get('testnet')
-        self.rawPrivKey = pycoin_wallet.secret_exponent
-        self.publicPoint = BasePoint * self.rawPrivKey
-        self.address = public_pair_to_bitcoin_address(self.publicPoint.pair(),
-                                                      compressed=False,
-                                                      is_test=self.testnet)
-
-
-class LooseAddressRecord(AddressRecord):
-    """Subclass of AddressRecord which is entirely imported.
-    The address may be an existing one.
-    """
-    def __init__(self, **kwargs):
-        """Create a LooseAddressRecord for a given wallet <model>,
-        color <color_set> and address <address_data>. Also let the constructor
-        know whether it's on <testnet> (optional).
-        <address_data> consists of:
-        privKey
-        pubKey
-        """
-        self.model = kwargs.get('model')
-        self.color_set = ColorSet(self.model.get_color_map(),
-                                  kwargs.get('color_set'))
-        self.testnet = kwargs.get('testnet')
-        self.rawPrivKey = from_bytes_32(
-            a2b_hashed_base58(kwargs['address_data']))
-        self.publicPoint = BasePoint * self.rawPrivKey
-        self.address = public_pair_to_bitcoin_address(self.publicPoint.pair(),
-                                                      compressed=False,
-                                                      is_test=self.testnet)
-
-
-class DWalletAddressManager(object):
-    """This class manages the creation of new AddressRecords.
-    Specifically, it keeps track of which colors have been created
-    in this wallet and how many addresses of each color have been
-    created in this wallet.
-    """
-    def __init__(self, model, config):
-        """Create a deterministic wallet address manager given
-        a wallet <model> and a configuration <config>.
-        Note address manager configuration is in the key "dwam".
-        """
-        self.config = config
-        self.testnet = config.get('testnet', False)
-        self.model = model
-        self.addresses = []
-
-        # initialize the wallet manager if this is the first time
-        #  this will generate a master key.
-        params = config.get('dwam', None)
-        if params is None:
-            params = self.init_new_wallet()
-
-        # master key is stored in a separate config entry
-        self.master_key = config['dw_master_key']
-
-        self.genesis_color_sets = params['genesis_color_sets']
-        self.color_set_states = params['color_set_states']
-
-        # import the genesis addresses
-        for i, color_desc_list in enumerate(self.genesis_color_sets):
-            addr = self.get_genesis_address(i)
-            addr.color_set = ColorSet(self.model.get_color_map(),
-                                      color_desc_list)
-            self.addresses.append(addr)
-
-        # now import the specific color addresses
-        for color_set_st in self.color_set_states:
-            color_desc_list = color_set_st['color_set']
-            max_index = color_set_st['max_index']
-            color_set = ColorSet(self.model.get_color_map(), color_desc_list)
-            params = {
-                'testnet': self.testnet,
-                'master_key': self.master_key,
-                'color_set': color_set
-                }
-            for index in xrange(max_index + 1):
-                params['index'] = index
-                self.addresses.append(DeterministicAddressRecord(**params))
-
-        # import the one-off addresses from the config
-        for addr_params in config.get('addresses', []):
-            addr_params['testnet'] = self.testnet
-            addr_params['model'] = model
-            try:
-                address = LooseAddressRecord(**addr_params)
-                self.addresses.append(address)
-            except InvalidAddressError:
-                address_type = "Testnet" if self.testnet else "Bitcoin"
-                #print "%s is an invalid %s address" % (
-                #    addr_params['address_data']['pubkey'], address_type)
-
-    def init_new_wallet(self):
-        """Initialize the configuration if this is the first time
-        we're creating addresses in this wallet.
-        Returns the "dwam" part of the configuration.
-        """
-        if not 'dw_master_key' in self.config:
-            master_key = os.urandom(64).encode('hex')
-            self.config['dw_master_key'] = master_key
-        dwam_params = {
-            'genesis_color_sets': [],
-            'color_set_states': []
-            }
-        self.config['dwam'] = dwam_params
-        return dwam_params
-
-    def increment_max_index_for_color_set(self, color_set):
-        """Given a color <color_set>, record that there is one more
-        new address for that color.
-        """
-        # TODO: speed up, cache(?)
-        for color_set_st in self.color_set_states:
-            color_desc_list = color_set_st['color_set']
-            max_index = color_set_st['max_index']
-            cur_color_set = ColorSet(self.model.get_color_map(),
-                                     color_desc_list)
-            if cur_color_set.equals(color_set):
-                max_index += 1
-                color_set_st['max_index'] = max_index
-                return max_index
-        self.color_set_states.append({"color_set": color_set.get_data(),
-                                      "max_index": 0})
-        return 0
-
-    def get_new_address(self, asset_or_color_set):
-        """Given an asset or color_set <asset_or_color_set>,
-        Create a new DeterministicAddressRecord and return it.
-        The DWalletAddressManager will keep that tally and
-        persist it in storage, so the address will be available later.
-        """
-        if isinstance(asset_or_color_set, AssetDefinition):
-            color_set = asset_or_color_set.get_color_set()
-        else:
-            color_set = asset_or_color_set
-        index = self.increment_max_index_for_color_set(color_set)
-        na = DeterministicAddressRecord(master_key=self.master_key,
-                                        color_set=color_set, index=index,
-                                        testnet=self.testnet)
-        self.addresses.append(na)
-        self.update_config()
-        return na
-
-    def get_genesis_address(self, genesis_index):
-        """Given the index <genesis_index>, will return
-        the Deterministic Address Record associated with that
-        index. In general, that index corresponds to the nth
-        color created by this wallet.
-        """
-        return DeterministicAddressRecord(
-            master_key=self.master_key,
-            color_set=ColorSet(self.model.get_color_map(), []),
-            index=genesis_index, testnet=self.testnet)
-
-    def get_new_genesis_address(self):
-        """Create a new genesis address and return it.
-        This will necessarily increment the number of genesis
-        addresses from this wallet.
-        """
-        index = len(self.genesis_color_sets)
-        self.genesis_color_sets.append([])
-        self.update_config()
-        address = self.get_genesis_address(index)
-        address.index = index
-        self.addresses.append(address)
-        return address
-
-    def update_genesis_address(self, address, color_set):
-        """Updates the genesis address <address> to have a different
-        color set <color_set>.
-        """
-        assert address.color_set.color_id_set == set([])
-        address.color_set = color_set
-        self.genesis_color_sets[address.index] = color_set.get_data()
-        self.update_config()
-
-    def get_some_address(self, color_set):
-        """Returns an address associated with color <color_set>.
-        This address will be essentially a random address in the
-        wallet. No guarantees to what will come out.
-        If there is not address corresponding to the color_set,
-        thhis method will create one and return it.
-        """
-        acs = self.get_addresses_for_color_set(color_set)
-        if acs:
-            # reuse
-            return acs[0]
-        else:
-            return self.get_new_address(color_set)
-
-    def get_change_address(self, color_set):
-        """Returns an address that can receive the change amount
-        for a color <color_set>
-        """
-        return self.get_some_address(color_set)
-
-    def get_all_addresses(self):
-        """Returns the list of all AddressRecords in this wallet.
-        """
-        return self.addresses
-
-    def get_addresses_for_color_set(self, color_set):
-        """Given a color <color_set>, returns all AddressRecords
-        that have that color.
-        """
-        return [addr for addr in self.addresses
-                if color_set.intersects(addr.get_color_set())]
-
-    def update_config(self):
-        """Updates the configuration for the address manager.
-        The data will persist in the key "dwam" and consists
-        of this data:
-        genesis_color_sets - Colors created by this wallet
-        color_set_states   - How many addresses of each color
-        """
-        dwam_params = {
-            'genesis_color_sets': self.genesis_color_sets,
-            'color_set_states': self.color_set_states
-            }
-        self.config['dwam'] = dwam_params
-
-
-class HDWalletAddressManager(DWalletAddressManager):
-    """This class manages the creation of new AddressRecords.
-    Specifically, it keeps track of which colors have been created
-    in this wallet and how many addresses of each color have been
-    created in this wallet. This is different from DWalletAddressManager
-    in that it is BIP-0032 compliant.
-    """
-    def __init__(self, model, config):
-        """Create a deterministic wallet address manager given
-        a wallet <model> and a configuration <config>.
-        Note address manager configuration is in the key "hdwam".
-        """
-        self.config = config
-        self.testnet = config.get('testnet', False)
-        self.model = model
-        self.addresses = []
-
-        # initialize the wallet manager if this is the first time
-        #  this will generate a master key.
-        params = config.get('hdwam', None)
-        if params is None:
-            params = self.init_new_wallet()
-
-        # master key is stored in a separate config entry
-        self.master_key = config['hdw_master_key']
-
-        master = hashlib.sha512(self.master_key.decode('hex')).digest()
-
-        # initialize a BIP-0032 wallet
-        self.pycoin_wallet = Wallet(is_private=True, is_test=self.testnet,
-                                    chain_code=master[32:],
-                                    secret_exponent_bytes=master[:32])
-
-        self.genesis_color_sets = params['genesis_color_sets']
-        self.color_set_states = params['color_set_states']
-
-        # import the genesis addresses
-        for i, color_desc_list in enumerate(self.genesis_color_sets):
-            addr = self.get_genesis_address(i)
-            addr.color_set = ColorSet(self.model.get_color_map(),
-                                      color_desc_list)
-            self.addresses.append(addr)
-
-        # now import the specific color addresses
-        for color_set_st in self.color_set_states:
-            color_desc_list = color_set_st['color_set']
-            max_index = color_set_st['max_index']
-            color_set = ColorSet(self.model.get_color_map(), color_desc_list)
-            params = {
-                'testnet': self.testnet,
-                'pycoin_wallet': self.pycoin_wallet,
-                'color_set': color_set
-                }
-            for index in xrange(max_index + 1):
-                params['index'] = index
-                self.addresses.append(BIP0032AddressRecord(**params))
-
-        # import the one-off addresses from the config
-        for addr_params in config.get('addresses', []):
-            addr_params['testnet'] = self.testnet
-            addr_params['model'] = model
-            try:
-                address = LooseAddressRecord(**addr_params)
-                self.addresses.append(address)
-            except InvalidAddressError:
-                address_type = "Testnet" if self.testnet else "Bitcoin"
-
-    def init_new_wallet(self):
-        """Initialize the configuration if this is the first time
-        we're creating addresses in this wallet.
-        Returns the "hdwam" part of the configuration.
-        """
-        if not 'hdw_master_key' in self.config:
-            master_key = os.urandom(64).encode('hex')
-            self.config['hdw_master_key'] = master_key
-        hdwam_params = {
-            'genesis_color_sets': [],
-            'color_set_states': []
-            }
-        self.config['hdwam'] = hdwam_params
-        return hdwam_params
-
-    def get_new_address(self, asset_or_color_set):
-        """Given an asset or color_set <asset_or_color_set>,
-        Create a new BIP0032AddressRecord and return it.
-        This class will keep that tally and
-        persist it in storage, so the address will be available later.
-        """
-        if isinstance(asset_or_color_set, AssetDefinition):
-            color_set = asset_or_color_set.get_color_set()
-        else:
-            color_set = asset_or_color_set
-        index = self.increment_max_index_for_color_set(color_set)
-        na = BIP0032AddressRecord(
-            pycoin_wallet=self.pycoin_wallet, color_set=color_set,
-            index=index, testnet=self.testnet)
-        self.addresses.append(na)
-        self.update_config()
-        return na
-
-    def get_genesis_address(self, genesis_index):
-        """Given the index <genesis_index>, will return
-        the BIP0032 Address Record associated with that
-        index. In general, that index corresponds to the nth
-        color created by this wallet.
-        """
-        return BIP0032AddressRecord(
-            pycoin_wallet=self.pycoin_wallet,
-            color_set=ColorSet(self.model.get_color_map(), []),
-            index=genesis_index, testnet=self.testnet)
-
-    def update_config(self):
-        """Updates the configuration for the address manager.
-        The data will persist in the key "dwam" and consists
-        of this data:
-        genesis_color_sets - Colors created by this wallet
-        color_set_states   - How many addresses of each color
-        """
-        dwam_params = {
-            'genesis_color_sets': self.genesis_color_sets,
-            'color_set_states': self.color_set_states
-            }
-        self.config['hdwam'] = dwam_params
 
 
 class ColoredCoinContext(object):
@@ -554,13 +106,15 @@ class WalletModel(object):
         self.ccc = ColoredCoinContext(config)
         self.ass_def_man = AssetDefinitionManager(self.ccc.colormap, config)
         if config.get('bip0032'):
-            self.address_man = HDWalletAddressManager(self, config)
+            from bip0032 import HDWalletAddressManager
+            self.address_man = HDWalletAddressManager(self.ccc.colormap, config)
         else:
-            self.address_man = DWalletAddressManager(self, config)
+            from deterministic import DWalletAddressManager
+            self.address_man = DWalletAddressManager(self.ccc.colormap, config)
 
         self.coin_query_factory = CoinQueryFactory(self, config)
         self.utxo_man = UTXOManager(self, config)
-        self.txdb = txdb.TxDb(self, config)
+        self.txdb = TxDb(self, config)
         self.testnet = config.get('testnet', False)
         self.tx_spec_transformer = TransactionSpecTransformer(self, config)
